create separate component for farm list in farm page and  APR for dragon syrup, reusable component for switch on farm/dragon page and update mobile view of farm/dragon cards
