+++ src/components/FarmDualCard/FarmDualCard.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 import { Box, Typography, useMediaQuery } from '@material-ui/core';
 import { makeStyles, useTheme } from '@material-ui/core/styles';
+import { KeyboardArrowDown, KeyboardArrowUp } from '@material-ui/icons';
 import { DualStakingInfo } from 'state/stake/hooks';
 import { JSBI, TokenAmount, ETHER } from '@uniswap/sdk';
 import { unwrappedToken } from 'utils/wrappedCurrency';
@@ -27,9 +28,6 @@ const useStyles = makeStyles(({ palette, breakpoints }) => ({
     alignItems: 'center',
     padding: '16px',
     cursor: 'pointer',
-    [breakpoints.down('xs')]: {
-      flexDirection: 'column',
-    },
   },
   farmDualText: {
     fontSize: 14,
@@ -131,120 +129,108 @@ const FarmDualCard: React.FC<{
     stakingInfo?.rateA * stakingInfo?.quickPrice +
     stakingInfo?.rateB * Number(stakingInfo.rewardTokenBPrice);
 
+  const renderPool = (width: number) => (
+    <Box display='flex' alignItems='center' width={width}>
+      <DoubleCurrencyLogo
+        currency0={currency0}
+        currency1={currency1}
+        size={28}
+      />
+      <Box ml={1.5}>
+        <Typography variant='body2'>
+          {currency0.symbol} / {currency1.symbol} LP
+        </Typography>
+      </Box>
+    </Box>
+  );
+
   return (
     <Box className={classes.farmDualCard}>
       <Box
         className={classes.farmDualCardUp}
         onClick={() => setExpandCard(!isExpandCard)}
       >
-        <Box
-          display='flex'
-          alignItems='center'
-          justifyContent='space-between'
-          width={isMobile ? 1 : 0.3}
-          mb={isMobile ? 1.5 : 0}
-        >
-          {isMobile && (
-            <Typography className={classes.farmDualText}>Pool</Typography>
-          )}
-          <Box display='flex' alignItems='center'>
-            <DoubleCurrencyLogo
-              currency0={currency0}
-              currency1={currency1}
-              size={28}
-            />
-            <Box ml={1.5}>
+        {isMobile ? (
+          <>
+            {renderPool(isExpandCard ? 0.95 : 0.7)}
+            {!isExpandCard && (
+              <Box width={0.25}>
+                <Box display='flex' alignItems='center'>
+                  <Typography variant='caption' color='textSecondary'>
+                    APY
+                  </Typography>
+                  <Box ml={0.5} height={16}>
+                    <img src={CircleInfoIcon} alt={'arrow up'} />
+                  </Box>
+                </Box>
+                <Box mt={0.5} color={palette.success.main}>
+                  <Typography variant='body2'>{apyWithFee}%</Typography>
+                </Box>
+              </Box>
+            )}
+            <Box
+              width={0.05}
+              display='flex'
+              justifyContent='flex-end'
+              color={palette.primary.main}
+            >
+              {isExpandCard ? <KeyboardArrowUp /> : <KeyboardArrowDown />}
+            </Box>
+          </>
+        ) : (
+          <>
+            {renderPool(0.3)}
+            <Box width={0.2} textAlign='center'>
+              <Typography variant='body2'>{tvl}</Typography>
+            </Box>
+            <Box width={0.25} textAlign='center'>
               <Typography variant='body2'>
-                {currency0.symbol} / {currency1.symbol} LP
+                ${Number(rewards.toFixed(0)).toLocaleString()} / day
               </Typography>
+              <Typography variant='body2'>{poolRateA}</Typography>
+              <Typography variant='body2'>{poolRateB}</Typography>
             </Box>
-          </Box>
-        </Box>
-        <Box
-          width={isMobile ? 1 : 0.2}
-          mb={isMobile ? 1.5 : 0}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-          alignItems='center'
-        >
-          {isMobile && (
-            <Typography className={classes.farmDualText}>TVL</Typography>
-          )}
-          <Typography variant='body2'>{tvl}</Typography>
-        </Box>
-        <Box
-          mb={isMobile ? 1.5 : 0}
-          width={isMobile ? 1 : 0.25}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-          alignItems='center'
-        >
-          {isMobile && (
-            <Typography className={classes.farmDualText}>Rewards</Typography>
-          )}
-          <Box textAlign={isMobile ? 'right' : 'left'}>
-            <Typography variant='body2'>{`$${parseInt(
-              rewards.toFixed(0),
-            ).toLocaleString()} / day`}</Typography>
-            <Typography variant='body2'>{poolRateA}</Typography>
-            <Typography variant='body2'>{poolRateB}</Typography>
-          </Box>
-        </Box>
-        <Box
-          mb={isMobile ? 1.5 : 0}
-          width={isMobile ? 1 : 0.15}
-          display='flex'
-          alignItems='center'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-        >
-          {isMobile && (
-            <Typography className={classes.farmDualText}>APY</Typography>
-          )}
-          <Box display='flex' alignItems='center'>
-            <Typography variant='body2' style={{ color: palette.success.main }}>
-              {apyWithFee}%
-            </Typography>
-            <Box ml={1} style={{ height: '16px' }}>
-              <img src={CircleInfoIcon} alt={'arrow up'} />
+            <Box
+              width={0.15}
+              display='flex'
+              justifyContent='center'
+              alignItems='center'
+              color={palette.success.main}
+            >
+              <Typography variant='body2'>{apyWithFee}%</Typography>
+              <Box ml={0.5} height={16}>
+                <img src={CircleInfoIcon} alt={'arrow up'} />
+              </Box>
             </Box>
-          </Box>
-        </Box>
-        <Box
-          width={isMobile ? 1 : 0.2}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'flex-end'}
-        >
-          {isMobile && (
-            <Typography className={classes.farmDualText}>Earned</Typography>
-          )}
-          <Box textAlign='right'>
-            <Typography variant='body2'>{earnedUSDStr}</Typography>
-            <Box display='flex' alignItems='center' justifyContent='flex-end'>
-              <CurrencyLogo
-                currency={returnTokenFromKey('QUICK')}
-                size='16px'
-              />
-              <Typography variant='body2' style={{ marginLeft: 5 }}>
-                {stakingInfo.earnedAmountA.toSignificant(2)}
-                <span>&nbsp;dQUICK</span>
-              </Typography>
+            <Box width={0.2} textAlign='right'>
+              <Typography variant='body2'>{earnedUSDStr}</Typography>
+              <Box display='flex' alignItems='center' justifyContent='flex-end'>
+                <CurrencyLogo
+                  currency={returnTokenFromKey('QUICK')}
+                  size='16px'
+                />
+                <Typography variant='body2' style={{ marginLeft: 5 }}>
+                  {stakingInfo.earnedAmountA.toSignificant(2)}
+                  <span>&nbsp;dQUICK</span>
+                </Typography>
+              </Box>
+              <Box display='flex' alignItems='center' justifyContent='flex-end'>
+                <CurrencyLogo
+                  currency={
+                    rewardTokenB.symbol?.toLowerCase() === 'wmatic'
+                      ? ETHER
+                      : rewardTokenB
+                  }
+                  size='16px'
+                />
+                <Typography variant='body2' style={{ marginLeft: 5 }}>
+                  {stakingInfo.earnedAmountB.toSignificant(2)}
+                  <span>&nbsp;{rewardTokenB.symbol}</span>
+                </Typography>
+              </Box>
             </Box>
-            <Box display='flex' alignItems='center' justifyContent='flex-end'>
-              <CurrencyLogo
-                currency={
-                  rewardTokenB.symbol?.toLowerCase() === 'wmatic'
-                    ? ETHER
-                    : rewardTokenB
-                }
-                size='16px'
-              />
-              <Typography variant='body2' style={{ marginLeft: 5 }}>
-                {stakingInfo.earnedAmountB.toSignificant(2)}
-                <span>&nbsp;{rewardTokenB.symbol}</span>
-              </Typography>
-            </Box>
-          </Box>
-        </Box>
+          </>
+        )}
       </Box>
 
       {isExpandCard && (
