+++ src/components/FarmLPCard/FarmLPCard.tsx
@@ -8,8 +8,9 @@ import { DoubleCurrencyLogo, CurrencyLogo } from 'components';
 import CircleInfoIcon from 'assets/images/circleinfo.svg';
 import FarmLPCardDetails from './FarmLPCardDetails';
 import { getAPYWithFee, returnTokenFromKey } from 'utils';
+import { KeyboardArrowDown, KeyboardArrowUp } from '@material-ui/icons';
 
-const useStyles = makeStyles(({ palette, breakpoints }) => ({
+const useStyles = makeStyles(({ palette }) => ({
   farmLPCard: {
     background: palette.secondary.dark,
     width: '100%',
@@ -25,11 +26,9 @@ const useStyles = makeStyles(({ palette, breakpoints }) => ({
     borderRadius: 10,
     display: 'flex',
     alignItems: 'center',
-    padding: '16px',
+    height: 74,
+    padding: '0 16px',
     cursor: 'pointer',
-    [breakpoints.down('xs')]: {
-      flexDirection: 'column',
-    },
   },
   farmLPText: {
     fontSize: 14,
@@ -121,105 +120,93 @@ const FarmLPCard: React.FC<{
 
   const rewards = stakingInfo?.dQuickToQuick * stakingInfo?.quickPrice;
 
+  const renderPool = (width: number) => (
+    <Box display='flex' alignItems='center' width={width}>
+      <DoubleCurrencyLogo
+        currency0={currency0}
+        currency1={currency1}
+        size={28}
+      />
+      <Box ml={1.5}>
+        <Typography variant='body2'>
+          {currency0.symbol} / {currency1.symbol} LP
+        </Typography>
+      </Box>
+    </Box>
+  );
+
   return (
     <Box className={classes.farmLPCard}>
       <Box
         className={classes.farmLPCardUp}
         onClick={() => setExpandCard(!isExpandCard)}
       >
-        <Box
-          display='flex'
-          alignItems='center'
-          justifyContent='space-between'
-          width={isMobile ? 1 : 0.3}
-          mb={isMobile ? 1.5 : 0}
-        >
-          {isMobile && (
-            <Typography className={classes.farmLPText}>Pool</Typography>
-          )}
-          <Box display='flex' alignItems='center'>
-            <DoubleCurrencyLogo
-              currency0={currency0}
-              currency1={currency1}
-              size={28}
-            />
-            <Box ml={1.5}>
+        {isMobile ? (
+          <>
+            {renderPool(isExpandCard ? 0.95 : 0.7)}
+            {!isExpandCard && (
+              <Box width={0.25}>
+                <Box display='flex' alignItems='center'>
+                  <Typography variant='caption' color='textSecondary'>
+                    APY
+                  </Typography>
+                  <Box ml={0.5} height={16}>
+                    <img src={CircleInfoIcon} alt={'arrow up'} />
+                  </Box>
+                </Box>
+                <Box mt={0.5} color={palette.success.main}>
+                  <Typography variant='body2'>{apyWithFee}%</Typography>
+                </Box>
+              </Box>
+            )}
+            <Box
+              width={0.05}
+              display='flex'
+              justifyContent='flex-end'
+              color={palette.primary.main}
+            >
+              {isExpandCard ? <KeyboardArrowUp /> : <KeyboardArrowDown />}
+            </Box>
+          </>
+        ) : (
+          <>
+            {renderPool(0.3)}
+            <Box width={0.2} textAlign='center'>
+              <Typography variant='body2'>{tvl}</Typography>
+            </Box>
+            <Box width={0.25} textAlign='center'>
               <Typography variant='body2'>
-                {currency0.symbol} / {currency1.symbol} LP
+                ${Number(rewards.toFixed(0)).toLocaleString()} / day
               </Typography>
+              <Typography variant='body2'>{poolRate}</Typography>
             </Box>
-          </Box>
-        </Box>
-        <Box
-          width={isMobile ? 1 : 0.2}
-          mb={isMobile ? 1.5 : 0}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-          alignItems='center'
-        >
-          {isMobile && (
-            <Typography className={classes.farmLPText}>TVL</Typography>
-          )}
-          <Typography variant='body2'>{tvl}</Typography>
-        </Box>
-        <Box
-          mb={isMobile ? 1.5 : 0}
-          width={isMobile ? 1 : 0.25}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-          alignItems='center'
-        >
-          {isMobile && (
-            <Typography className={classes.farmLPText}>Rewards</Typography>
-          )}
-          <Box textAlign={isMobile ? 'right' : 'left'}>
-            <Typography variant='body2'>
-              ${Number(rewards.toFixed(0)).toLocaleString()} / day
-            </Typography>
-            <Typography variant='body2'>{poolRate}</Typography>
-          </Box>
-        </Box>
-        <Box
-          mb={isMobile ? 1.5 : 0}
-          width={isMobile ? 1 : 0.15}
-          display='flex'
-          alignItems='center'
-          justifyContent={isMobile ? 'space-between' : 'center'}
-        >
-          {isMobile && (
-            <Typography className={classes.farmLPText}>APY</Typography>
-          )}
-          <Box display='flex' alignItems='center'>
-            <Typography variant='body2' style={{ color: palette.success.main }}>
-              {apyWithFee}%
-            </Typography>
-            <Box ml={1} style={{ height: '16px' }}>
-              <img src={CircleInfoIcon} alt={'arrow up'} />
+            <Box
+              width={0.15}
+              display='flex'
+              justifyContent='center'
+              alignItems='center'
+              color={palette.success.main}
+            >
+              <Typography variant='body2'>{apyWithFee}%</Typography>
+              <Box ml={0.5} height={16}>
+                <img src={CircleInfoIcon} alt={'arrow up'} />
+              </Box>
             </Box>
-          </Box>
-        </Box>
-        <Box
-          width={isMobile ? 1 : 0.2}
-          display='flex'
-          justifyContent={isMobile ? 'space-between' : 'flex-end'}
-        >
-          {isMobile && (
-            <Typography className={classes.farmLPText}>Earned</Typography>
-          )}
-          <Box textAlign='right'>
-            <Typography variant='body2'>{earnedUSDStr}</Typography>
-            <Box display='flex' alignItems='center' justifyContent='flex-end'>
-              <CurrencyLogo
-                currency={returnTokenFromKey('QUICK')}
-                size='16px'
-              />
-              <Typography variant='body2' style={{ marginLeft: 5 }}>
-                {stakingInfo.earnedAmount.toSignificant(2)}
-                <span>&nbsp;dQUICK</span>
-              </Typography>
+            <Box width={0.2} textAlign='right'>
+              <Typography variant='body2'>{earnedUSDStr}</Typography>
+              <Box display='flex' alignItems='center' justifyContent='flex-end'>
+                <CurrencyLogo
+                  currency={returnTokenFromKey('QUICK')}
+                  size='16px'
+                />
+                <Typography variant='body2' style={{ marginLeft: 5 }}>
+                  {stakingInfo.earnedAmount.toSignificant(2)}
+                  <span>&nbsp;dQUICK</span>
+                </Typography>
+              </Box>
             </Box>
-          </Box>
-        </Box>
+          </>
+        )}
       </Box>
 
       {isExpandCard && (
