+++ src/components/FarmLPCard/FarmLPCardDetails.tsx
@@ -15,8 +15,9 @@ import { useActiveWeb3React } from 'hooks';
 import useTransactionDeadline from 'hooks/useTransactionDeadline';
 import { useApproveCallback, ApprovalState } from 'hooks/useApproveCallback';
 import { getAPYWithFee, returnTokenFromKey } from 'utils';
+import CircleInfoIcon from 'assets/images/circleinfo.svg';
 
-const useStyles = makeStyles(({ palette }) => ({
+const useStyles = makeStyles(({ palette, breakpoints }) => ({
   inputVal: {
     backgroundColor: palette.secondary.contrastText,
     borderRadius: '10px',
@@ -59,6 +60,17 @@ const useStyles = makeStyles(({ palette }) => ({
     justifyContent: 'center',
     cursor: 'pointer',
     color: 'white',
+  },
+  buttonWrapper: {
+    minWidth: 250,
+    width: '30%',
+    color: palette.text.secondary,
+    borderRadius: 16,
+    [breakpoints.down('xs')]: {
+      width: '100%',
+      padding: 16,
+      border: `1px solid ${palette.divider}`,
+    },
   },
 }));
 
@@ -114,7 +126,23 @@ const FarmLPCardDetails: React.FC<{
 
   let valueOfMyStakedAmountInBaseToken: TokenAmount | undefined;
   let valueOfUnstakedAmountInBaseToken: TokenAmount | undefined;
+  let valueOfTotalStakedAmountInBaseToken: TokenAmount | undefined;
   if (stakingInfo && stakingInfo.totalSupply && pair && baseToken) {
+    // take the total amount of LP tokens staked, multiply by ETH value of all LP tokens, divide by all LP tokens
+    valueOfTotalStakedAmountInBaseToken = new TokenAmount(
+      baseToken,
+      JSBI.divide(
+        JSBI.multiply(
+          JSBI.multiply(
+            stakingInfo.totalStakedAmount.raw,
+            pair.reserveOf(baseToken).raw,
+          ),
+          JSBI.BigInt(2), // this is b/c the value of LP shares are ~double the value of the WETH they entitle owner to
+        ),
+        stakingInfo.totalSupply.raw,
+      ),
+    );
+
     valueOfMyStakedAmountInBaseToken = new TokenAmount(
       baseToken,
       JSBI.divide(
@@ -306,163 +334,201 @@ const FarmLPCardDetails: React.FC<{
       ? '< $0.001'
       : '$' + earnedUSD.toLocaleString();
 
+  const valueOfTotalStakedAmountInUSDC =
+    valueOfTotalStakedAmountInBaseToken &&
+    USDPrice?.quote(valueOfTotalStakedAmountInBaseToken);
+
+  const tvl = valueOfTotalStakedAmountInUSDC
+    ? `$${valueOfTotalStakedAmountInUSDC.toFixed(0, { groupSeparator: ',' })}`
+    : `${valueOfTotalStakedAmountInBaseToken?.toSignificant(4, {
+        groupSeparator: ',',
+      }) ?? '-'} ETH`;
+
+  const rewards =
+    (stakingInfo?.dQuickToQuick ?? 0) * (stakingInfo?.quickPrice ?? 0);
+
+  const poolRate = `${stakingInfo?.totalRewardRate
+    ?.toFixed(2, { groupSeparator: ',' })
+    .replace(/[.,]00$/, '')} dQUICK / day`;
+
   return (
     <Box
       width='100%'
-      mt={2.5}
-      pl={isMobile ? 2 : 4}
-      pr={isMobile ? 2 : 4}
-      pt={2}
+      p={2}
       display='flex'
       flexDirection='row'
       flexWrap='wrap'
       borderTop='1px solid #444444'
       alignItems='center'
-      justifyContent='space-between'
+      justifyContent={stakingInfo?.ended ? 'flex-end' : 'space-between'}
     >
       {stakingInfo && (
         <>
-          <Box
-            minWidth={250}
-            width={isMobile ? 1 : 0.3}
-            color={palette.text.secondary}
-            my={stakingInfo.ended ? 0 : 1.5}
-          >
-            {!stakingInfo.ended && (
-              <>
-                <Box
-                  display='flex'
-                  flexDirection='row'
-                  alignItems='flex-start'
-                  justifyContent='space-between'
-                >
-                  <Typography variant='body2'>In Wallet:</Typography>
-                  <Box
-                    display='flex'
-                    flexDirection='column'
-                    alignItems='flex-end'
-                    justifyContent='flex-start'
-                  >
-                    <Typography variant='body2'>
-                      {userLiquidityUnstaked
-                        ? userLiquidityUnstaked.toSignificant(2)
-                        : 0}{' '}
-                      LP{' '}
-                      <span>
-                        (
-                        {valueOfUnstakedAmountInUSDC
-                          ? Number(
-                              valueOfUnstakedAmountInUSDC.toSignificant(2),
-                            ) > 0 &&
-                            Number(
-                              valueOfUnstakedAmountInUSDC.toSignificant(2),
-                            ) < 0.001
-                            ? '< $0.001'
-                            : `$${valueOfUnstakedAmountInUSDC.toSignificant(2)}`
-                          : '$0'}
-                        )
-                      </span>
-                    </Typography>
-                    <Link
-                      to={`/pools?currency0=${
-                        token0?.symbol?.toLowerCase() === 'wmatic'
-                          ? 'ETH'
-                          : token0?.address
-                      }&currency1=${
-                        token1?.symbol?.toLowerCase() === 'wmatic'
-                          ? 'ETH'
-                          : token1?.address
-                      }`}
-                      style={{ color: palette.primary.main }}
-                    >
-                      Get {currency0?.symbol} / {currency1?.symbol} LP
-                    </Link>
+          {isMobile && (
+            <>
+              <Box
+                mt={2}
+                width={1}
+                display='flex'
+                justifyContent='space-between'
+              >
+                <Typography variant='body2' color='textSecondary'>
+                  TVL
+                </Typography>
+                <Typography variant='body2'>{tvl}</Typography>
+              </Box>
+              <Box
+                mt={2}
+                width={1}
+                display='flex'
+                justifyContent='space-between'
+              >
+                <Typography variant='body2' color='textSecondary'>
+                  Rewards
+                </Typography>
+                <Box textAlign='right'>
+                  <Typography variant='body2'>
+                    ${Number(rewards.toFixed(0)).toLocaleString()} / day
+                  </Typography>
+                  <Typography variant='body2'>{poolRate}</Typography>
+                </Box>
+              </Box>
+              <Box
+                mt={2}
+                width={1}
+                display='flex'
+                justifyContent='space-between'
+              >
+                <Box display='flex' alignItems='center'>
+                  <Typography variant='body2' color='textSecondary'>
+                    APY
+                  </Typography>
+                  <Box ml={0.5} height={16}>
+                    <img src={CircleInfoIcon} alt={'arrow up'} />
                   </Box>
                 </Box>
-                <Box className={classes.inputVal} mb={2} mt={2} p={2}>
-                  <input
-                    placeholder='0.00'
-                    value={stakeAmount}
-                    onChange={(evt: any) => {
-                      setStakeAmount(evt.target.value);
-                    }}
-                  />
-                  <Typography
-                    variant='body2'
-                    style={{
-                      color:
-                        userLiquidityUnstaked &&
-                        userLiquidityUnstaked.greaterThan('0')
-                          ? palette.primary.main
-                          : palette.text.hint,
-                    }}
-                    onClick={() => {
-                      if (
-                        userLiquidityUnstaked &&
-                        userLiquidityUnstaked.greaterThan('0')
-                      ) {
-                        setStakeAmount(userLiquidityUnstaked.toExact());
-                      } else {
-                        setStakeAmount('');
-                      }
-                    }}
-                  >
-                    MAX
-                  </Typography>
+                <Box color={palette.success.main}>
+                  <Typography variant='body2'>{apyWithFee}%</Typography>
                 </Box>
+              </Box>
+            </>
+          )}
+          {!stakingInfo.ended && (
+            <Box className={classes.buttonWrapper} mt={isMobile ? 2 : 0}>
+              <Box display='flex' justifyContent='space-between'>
+                <Typography variant='body2'>In Wallet:</Typography>
                 <Box
-                  className={
-                    !approving &&
-                    Number(!attemptStaking && stakeAmount) > 0 &&
-                    Number(stakeAmount) <=
-                      Number(userLiquidityUnstaked?.toExact())
-                      ? classes.buttonClaim
-                      : classes.buttonToken
-                  }
-                  mb={2}
-                  mt={2}
-                  p={2}
-                  onClick={async () => {
+                  display='flex'
+                  flexDirection='column'
+                  alignItems='flex-end'
+                  justifyContent='flex-start'
+                >
+                  <Typography variant='body2'>
+                    {userLiquidityUnstaked
+                      ? userLiquidityUnstaked.toSignificant(2)
+                      : 0}{' '}
+                    LP{' '}
+                    <span>
+                      (
+                      {valueOfUnstakedAmountInUSDC
+                        ? Number(valueOfUnstakedAmountInUSDC.toSignificant(2)) >
+                            0 &&
+                          Number(valueOfUnstakedAmountInUSDC.toSignificant(2)) <
+                            0.001
+                          ? '< $0.001'
+                          : `$${valueOfUnstakedAmountInUSDC.toSignificant(2)}`
+                        : '$0'}
+                      )
+                    </span>
+                  </Typography>
+                  <Link
+                    to={`/pools?currency0=${
+                      token0?.symbol?.toLowerCase() === 'wmatic'
+                        ? 'ETH'
+                        : token0?.address
+                    }&currency1=${
+                      token1?.symbol?.toLowerCase() === 'wmatic'
+                        ? 'ETH'
+                        : token1?.address
+                    }`}
+                    style={{ color: palette.primary.main }}
+                  >
+                    Get {currency0?.symbol} / {currency1?.symbol} LP
+                  </Link>
+                </Box>
+              </Box>
+              <Box className={classes.inputVal} mb={2} mt={2} p={2}>
+                <input
+                  placeholder='0.00'
+                  value={stakeAmount}
+                  onChange={(evt: any) => {
+                    setStakeAmount(evt.target.value);
+                  }}
+                />
+                <Typography
+                  variant='body2'
+                  style={{
+                    color:
+                      userLiquidityUnstaked &&
+                      userLiquidityUnstaked.greaterThan('0')
+                        ? palette.primary.main
+                        : palette.text.hint,
+                  }}
+                  onClick={() => {
                     if (
-                      !approving &&
-                      !attemptStaking &&
-                      Number(stakeAmount) > 0 &&
-                      Number(stakeAmount) <=
-                        Number(userLiquidityUnstaked?.toExact())
+                      userLiquidityUnstaked &&
+                      userLiquidityUnstaked.greaterThan('0')
                     ) {
-                      if (approval === ApprovalState.APPROVED) {
-                        onStake();
-                      } else {
-                        onAttemptToApprove();
-                      }
+                      setStakeAmount(userLiquidityUnstaked.toExact());
+                    } else {
+                      setStakeAmount('');
                     }
                   }}
                 >
-                  <Typography variant='body1'>
-                    {attemptStaking
-                      ? 'Staking LP Tokens...'
-                      : approval === ApprovalState.APPROVED
-                      ? 'Stake LP Tokens'
-                      : approving
-                      ? 'Approving...'
-                      : 'Approve'}
-                  </Typography>
-                </Box>
-              </>
-            )}
-          </Box>
-          <Box
-            minWidth={250}
-            width={isMobile ? 1 : 0.3}
-            my={stakingInfo.ended ? 0 : 1.5}
-            color={palette.text.secondary}
-          >
-            <Box
-              display='flex'
-              flexDirection='row'
-              alignItems='flex-start'
-              justifyContent='space-between'
-            >
+                  MAX
+                </Typography>
+              </Box>
+              <Box
+                className={
+                  !approving &&
+                  Number(!attemptStaking && stakeAmount) > 0 &&
+                  Number(stakeAmount) <=
+                    Number(userLiquidityUnstaked?.toExact())
+                    ? classes.buttonClaim
+                    : classes.buttonToken
+                }
+                mt={2}
+                p={2}
+                onClick={async () => {
+                  if (
+                    !approving &&
+                    !attemptStaking &&
+                    Number(stakeAmount) > 0 &&
+                    Number(stakeAmount) <=
+                      Number(userLiquidityUnstaked?.toExact())
+                  ) {
+                    if (approval === ApprovalState.APPROVED) {
+                      onStake();
+                    } else {
+                      onAttemptToApprove();
+                    }
+                  }
+                }}
+              >
+                <Typography variant='body1'>
+                  {attemptStaking
+                    ? 'Staking LP Tokens...'
+                    : approval === ApprovalState.APPROVED
+                    ? 'Stake LP Tokens'
+                    : approving
+                    ? 'Approving...'
+                    : 'Approve'}
+                </Typography>
+              </Box>
+            </Box>
+          )}
+          <Box className={classes.buttonWrapper} mx={isMobile ? 0 : 2} my={2}>
+            <Box display='flex' justifyContent='space-between'>
               <Typography variant='body2'>My deposits:</Typography>
               <Typography variant='body2'>
                 {stakingInfo.stakedAmount.toSignificant(2)} LP{' '}
@@ -520,7 +586,6 @@ const FarmLPCardDetails: React.FC<{
                   ? classes.buttonClaim
                   : classes.buttonToken
               }
-              mb={2}
               mt={2}
               p={2}
               onClick={() => {
@@ -541,12 +606,7 @@ const FarmLPCardDetails: React.FC<{
               </Typography>
             </Box>
           </Box>
-          <Box
-            minWidth={250}
-            my={1.5}
-            width={isMobile ? 1 : 0.3}
-            color={palette.text.secondary}
-          >
+          <Box className={classes.buttonWrapper}>
             <Box
               display='flex'
               flexDirection='column'
@@ -575,7 +635,6 @@ const FarmLPCardDetails: React.FC<{
                   ? classes.buttonClaim
                   : classes.buttonToken
               }
-              mb={2}
               p={2}
               onClick={() => {
                 if (
